Analytics/pyproject.toml
[tool.poetry]
name = "myvote_analytics"
version = "0.1.0"
description = ""
authors = [
    'ИИрокезы'
]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
gunicorn = "^21.2.0"
pydantic = "^2"
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
asyncpg = {version = "^0.28.0", extras = ["sa"]}
redis = {version = "^4.6.0", extras = ["hiredis"]}
aiofiles = "^23.1.0"
httptools = "^0.6.0"
aio-pika = "^9.1.4"
profanity-check = "^1.0.3"
pyspellchecker = "^0.7.2"
keybert = "^0.7.0"
torch = {version = "^2.0.1+cpu", source = "pytorch-cpu-src"}
torchvision = {version = "^0.15.2+cpu", source = "pytorch-cpu-src"}


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
httpx = "^0.23.3"

[[tool.poetry.source]]
name = "pytorch-cpu-src"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["myvote_analytics",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "MYVOTE_ANALYTICS_ENVIRONMENT=pytest",
    "MYVOTE_ANALYTICS_DB_BASE=myvote_analytics_test",
]

[fastapi-template.options]
project_name = "myvote_analytics"
api_type = "rest"
enable_redis = "True"
enable_rmq = "True"
ci_type = "none"
enable_migrations = "None"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "myvote-analytics"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "None"
traefik_labels = "None"
add_dummy = "True"
orm = "sqlalchemy"
self_hosted_swagger = "True"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "True"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
